Type Definition atsamd51p::ccl::lutctrl::INSEL1_R

type INSEL1_R = R<u8, INSEL1_A>;

Reader of field INSEL1

Implementations

impl INSEL1_R

pub fn variant(&self) -> Variant<u8, INSEL1_A>

Get enumerated values variant

pub fn is_mask(&self) -> bool

Checks if the value of the field is MASK

pub fn is_feedback(&self) -> bool

Checks if the value of the field is FEEDBACK

pub fn is_link(&self) -> bool

Checks if the value of the field is LINK

pub fn is_event(&self) -> bool

Checks if the value of the field is EVENT

pub fn is_io(&self) -> bool

Checks if the value of the field is IO

pub fn is_ac(&self) -> bool

Checks if the value of the field is AC

pub fn is_tc(&self) -> bool

Checks if the value of the field is TC

pub fn is_alttc(&self) -> bool

Checks if the value of the field is ALTTC

pub fn is_tcc(&self) -> bool

Checks if the value of the field is TCC

pub fn is_sercom(&self) -> bool

Checks if the value of the field is SERCOM