Type Definition atsamd51p::dmac::channel::chctrla::BURSTLEN_R

type BURSTLEN_R = R<u8, BURSTLEN_A>;

Reader of field BURSTLEN

Implementations

impl BURSTLEN_R

pub fn variant(&self) -> BURSTLEN_A

Get enumerated values variant

pub fn is_single(&self) -> bool

Checks if the value of the field is SINGLE

pub fn is_2beat(&self) -> bool

Checks if the value of the field is _2BEAT

pub fn is_3beat(&self) -> bool

Checks if the value of the field is _3BEAT

pub fn is_4beat(&self) -> bool

Checks if the value of the field is _4BEAT

pub fn is_5beat(&self) -> bool

Checks if the value of the field is _5BEAT

pub fn is_6beat(&self) -> bool

Checks if the value of the field is _6BEAT

pub fn is_7beat(&self) -> bool

Checks if the value of the field is _7BEAT

pub fn is_8beat(&self) -> bool

Checks if the value of the field is _8BEAT

pub fn is_9beat(&self) -> bool

Checks if the value of the field is _9BEAT

pub fn is_10beat(&self) -> bool

Checks if the value of the field is _10BEAT

pub fn is_11beat(&self) -> bool

Checks if the value of the field is _11BEAT

pub fn is_12beat(&self) -> bool

Checks if the value of the field is _12BEAT

pub fn is_13beat(&self) -> bool

Checks if the value of the field is _13BEAT

pub fn is_14beat(&self) -> bool

Checks if the value of the field is _14BEAT

pub fn is_15beat(&self) -> bool

Checks if the value of the field is _15BEAT

pub fn is_16beat(&self) -> bool

Checks if the value of the field is _16BEAT