Type Definition atsamd51p::dmac::channel::chctrla::TRIGSRC_R

type TRIGSRC_R = R<u8, TRIGSRC_A>;

Reader of field TRIGSRC

Implementations

impl TRIGSRC_R

pub fn variant(&self) -> Variant<u8, TRIGSRC_A>

Get enumerated values variant

pub fn is_disable(&self) -> bool

Checks if the value of the field is DISABLE

pub fn is_rtc_timestamp(&self) -> bool

Checks if the value of the field is RTC_TIMESTAMP

pub fn is_dsu_dcc0(&self) -> bool

Checks if the value of the field is DSU_DCC0

pub fn is_dsu_dcc1(&self) -> bool

Checks if the value of the field is DSU_DCC1

pub fn is_sercom0_rx(&self) -> bool

Checks if the value of the field is SERCOM0_RX

pub fn is_sercom0_tx(&self) -> bool

Checks if the value of the field is SERCOM0_TX

pub fn is_sercom1_rx(&self) -> bool

Checks if the value of the field is SERCOM1_RX

pub fn is_sercom1_tx(&self) -> bool

Checks if the value of the field is SERCOM1_TX

pub fn is_sercom2_rx(&self) -> bool

Checks if the value of the field is SERCOM2_RX

pub fn is_sercom2_tx(&self) -> bool

Checks if the value of the field is SERCOM2_TX

pub fn is_sercom3_rx(&self) -> bool

Checks if the value of the field is SERCOM3_RX

pub fn is_sercom3_tx(&self) -> bool

Checks if the value of the field is SERCOM3_TX

pub fn is_sercom4_rx(&self) -> bool

Checks if the value of the field is SERCOM4_RX

pub fn is_sercom4_tx(&self) -> bool

Checks if the value of the field is SERCOM4_TX

pub fn is_sercom5_rx(&self) -> bool

Checks if the value of the field is SERCOM5_RX

pub fn is_sercom5_tx(&self) -> bool

Checks if the value of the field is SERCOM5_TX

pub fn is_sercom6_rx(&self) -> bool

Checks if the value of the field is SERCOM6_RX

pub fn is_sercom6_tx(&self) -> bool

Checks if the value of the field is SERCOM6_TX

pub fn is_sercom7_rx(&self) -> bool

Checks if the value of the field is SERCOM7_RX

pub fn is_sercom7_tx(&self) -> bool

Checks if the value of the field is SERCOM7_TX

pub fn is_can0_debug(&self) -> bool

Checks if the value of the field is CAN0_DEBUG

pub fn is_can1_debug(&self) -> bool

Checks if the value of the field is CAN1_DEBUG

pub fn is_tcc0_ovf(&self) -> bool

Checks if the value of the field is TCC0_OVF

pub fn is_tcc0_mc_0(&self) -> bool

Checks if the value of the field is TCC0_MC_0

pub fn is_tcc0_mc_1(&self) -> bool

Checks if the value of the field is TCC0_MC_1

pub fn is_tcc0_mc_2(&self) -> bool

Checks if the value of the field is TCC0_MC_2

pub fn is_tcc0_mc_3(&self) -> bool

Checks if the value of the field is TCC0_MC_3

pub fn is_tcc0_mc_4(&self) -> bool

Checks if the value of the field is TCC0_MC_4

pub fn is_tcc0_mc_5(&self) -> bool

Checks if the value of the field is TCC0_MC_5

pub fn is_tcc1_ovf(&self) -> bool

Checks if the value of the field is TCC1_OVF

pub fn is_tcc1_mc_0(&self) -> bool

Checks if the value of the field is TCC1_MC_0

pub fn is_tcc1_mc_1(&self) -> bool

Checks if the value of the field is TCC1_MC_1

pub fn is_tcc1_mc_2(&self) -> bool

Checks if the value of the field is TCC1_MC_2

pub fn is_tcc1_mc_3(&self) -> bool

Checks if the value of the field is TCC1_MC_3

pub fn is_tcc2_ovf(&self) -> bool

Checks if the value of the field is TCC2_OVF

pub fn is_tcc2_mc_0(&self) -> bool

Checks if the value of the field is TCC2_MC_0

pub fn is_tcc2_mc_1(&self) -> bool

Checks if the value of the field is TCC2_MC_1

pub fn is_tcc2_mc_2(&self) -> bool

Checks if the value of the field is TCC2_MC_2

pub fn is_tcc3_ovf(&self) -> bool

Checks if the value of the field is TCC3_OVF

pub fn is_tcc3_mc_0(&self) -> bool

Checks if the value of the field is TCC3_MC_0

pub fn is_tcc3_mc_1(&self) -> bool

Checks if the value of the field is TCC3_MC_1

pub fn is_tcc4_ovf(&self) -> bool

Checks if the value of the field is TCC4_OVF

pub fn is_tcc4_mc_0(&self) -> bool

Checks if the value of the field is TCC4_MC_0

pub fn is_tcc4_mc_1(&self) -> bool

Checks if the value of the field is TCC4_MC_1

pub fn is_tc0_ovf(&self) -> bool

Checks if the value of the field is TC0_OVF

pub fn is_tc0_mc_0(&self) -> bool

Checks if the value of the field is TC0_MC_0

pub fn is_tc0_mc_1(&self) -> bool

Checks if the value of the field is TC0_MC_1

pub fn is_tc1_ovf(&self) -> bool

Checks if the value of the field is TC1_OVF

pub fn is_tc1_mc_0(&self) -> bool

Checks if the value of the field is TC1_MC_0

pub fn is_tc1_mc_1(&self) -> bool

Checks if the value of the field is TC1_MC_1

pub fn is_tc2_ovf(&self) -> bool

Checks if the value of the field is TC2_OVF

pub fn is_tc2_mc_0(&self) -> bool

Checks if the value of the field is TC2_MC_0

pub fn is_tc2_mc_1(&self) -> bool

Checks if the value of the field is TC2_MC_1

pub fn is_tc3_ovf(&self) -> bool

Checks if the value of the field is TC3_OVF

pub fn is_tc3_mc_0(&self) -> bool

Checks if the value of the field is TC3_MC_0

pub fn is_tc3_mc_1(&self) -> bool

Checks if the value of the field is TC3_MC_1

pub fn is_tc4_ovf(&self) -> bool

Checks if the value of the field is TC4_OVF

pub fn is_tc4_mc_0(&self) -> bool

Checks if the value of the field is TC4_MC_0

pub fn is_tc4_mc_1(&self) -> bool

Checks if the value of the field is TC4_MC_1

pub fn is_tc5_ovf(&self) -> bool

Checks if the value of the field is TC5_OVF

pub fn is_tc5_mc_0(&self) -> bool

Checks if the value of the field is TC5_MC_0

pub fn is_tc5_mc_1(&self) -> bool

Checks if the value of the field is TC5_MC_1

pub fn is_tc6_ovf(&self) -> bool

Checks if the value of the field is TC6_OVF

pub fn is_tc6_mc_0(&self) -> bool

Checks if the value of the field is TC6_MC_0

pub fn is_tc6_mc_1(&self) -> bool

Checks if the value of the field is TC6_MC_1

pub fn is_tc7_ovf(&self) -> bool

Checks if the value of the field is TC7_OVF

pub fn is_tc7_mc_0(&self) -> bool

Checks if the value of the field is TC7_MC_0

pub fn is_tc7_mc_1(&self) -> bool

Checks if the value of the field is TC7_MC_1

pub fn is_adc0_resrdy(&self) -> bool

Checks if the value of the field is ADC0_RESRDY

pub fn is_adc0_seq(&self) -> bool

Checks if the value of the field is ADC0_SEQ

pub fn is_adc1_resrdy(&self) -> bool

Checks if the value of the field is ADC1_RESRDY

pub fn is_adc1_seq(&self) -> bool

Checks if the value of the field is ADC1_SEQ

pub fn is_dac_empty_0(&self) -> bool

Checks if the value of the field is DAC_EMPTY_0

pub fn is_dac_empty_1(&self) -> bool

Checks if the value of the field is DAC_EMPTY_1

pub fn is_dac_resrdy_0(&self) -> bool

Checks if the value of the field is DAC_RESRDY_0

pub fn is_dac_resrdy_1(&self) -> bool

Checks if the value of the field is DAC_RESRDY_1

pub fn is_i2s_rx_0(&self) -> bool

Checks if the value of the field is I2S_RX_0

pub fn is_i2s_rx_1(&self) -> bool

Checks if the value of the field is I2S_RX_1

pub fn is_i2s_tx_0(&self) -> bool

Checks if the value of the field is I2S_TX_0

pub fn is_i2s_tx_1(&self) -> bool

Checks if the value of the field is I2S_TX_1

pub fn is_pcc_rx(&self) -> bool

Checks if the value of the field is PCC_RX

pub fn is_aes_wr(&self) -> bool

Checks if the value of the field is AES_WR

pub fn is_aes_rd(&self) -> bool

Checks if the value of the field is AES_RD

pub fn is_qspi_rx(&self) -> bool

Checks if the value of the field is QSPI_RX

pub fn is_qspi_tx(&self) -> bool

Checks if the value of the field is QSPI_TX