Type Definition atsamd51p::i2s::rxctrl::DATASIZE_R

type DATASIZE_R = R<u8, DATASIZE_A>;

Reader of field DATASIZE

Implementations

impl DATASIZE_R

pub fn variant(&self) -> DATASIZE_A

Get enumerated values variant

pub fn is_32(&self) -> bool

Checks if the value of the field is _32

pub fn is_24(&self) -> bool

Checks if the value of the field is _24

pub fn is_20(&self) -> bool

Checks if the value of the field is _20

pub fn is_18(&self) -> bool

Checks if the value of the field is _18

pub fn is_16(&self) -> bool

Checks if the value of the field is _16

pub fn is_16c(&self) -> bool

Checks if the value of the field is _16C

pub fn is_8(&self) -> bool

Checks if the value of the field is _8

pub fn is_8c(&self) -> bool

Checks if the value of the field is _8C