atsamd51p::nvmctrl::seecfg

Type Alias WmodeW

pub type WmodeW<'a, REG> = BitWriter<'a, REG, Wmodeselect>;

Field WMODE writer - Write Mode

Aliased Type

struct WmodeW<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> WmodeW<'a, REG>
where REG: Writable + RegisterSpec,

pub fn unbuffered(self) -> &'a mut W<REG>

A NVM write command is issued after each write in the pagebuffer

pub fn buffered(self) -> &'a mut W<REG>

A NVM write command is issued when a write to a new page is requested