Type Definition atsamd51p::osc32kctrl::xosc32k::STARTUP_R

type STARTUP_R = R<u8, STARTUP_A>;

Reader of field STARTUP

Implementations

impl STARTUP_R

pub fn variant(&self) -> Variant<u8, STARTUP_A>

Get enumerated values variant

pub fn is_cycle2048(&self) -> bool

Checks if the value of the field is CYCLE2048

pub fn is_cycle4096(&self) -> bool

Checks if the value of the field is CYCLE4096

pub fn is_cycle16384(&self) -> bool

Checks if the value of the field is CYCLE16384

pub fn is_cycle32768(&self) -> bool

Checks if the value of the field is CYCLE32768

pub fn is_cycle65536(&self) -> bool

Checks if the value of the field is CYCLE65536

pub fn is_cycle131072(&self) -> bool

Checks if the value of the field is CYCLE131072

pub fn is_cycle262144(&self) -> bool

Checks if the value of the field is CYCLE262144