Type Definition atsamd51p::pac::wrctrl::KEY_R

type KEY_R = R<u8, KEY_A>;

Reader of field KEY

Implementations

impl KEY_R

pub fn variant(&self) -> Variant<u8, KEY_A>

Get enumerated values variant

pub fn is_off(&self) -> bool

Checks if the value of the field is OFF

pub fn is_clr(&self) -> bool

Checks if the value of the field is CLR

pub fn is_set(&self) -> bool

Checks if the value of the field is SET

pub fn is_setlck(&self) -> bool

Checks if the value of the field is SETLCK