Type Definition atsamd51p::pdec::evctrl::EVACT_R

type EVACT_R = R<u8, EVACT_A>;

Reader of field EVACT

Implementations

impl EVACT_R

pub fn variant(&self) -> Variant<u8, EVACT_A>

Get enumerated values variant

pub fn is_off(&self) -> bool

Checks if the value of the field is OFF

pub fn is_retrigger(&self) -> bool

Checks if the value of the field is RETRIGGER

pub fn is_count(&self) -> bool

Checks if the value of the field is COUNT