Type Definition atsamd51p::pm::hibcfg::RAMCFG_R

type RAMCFG_R = R<u8, RAMCFG_A>;

Reader of field RAMCFG

Implementations

impl RAMCFG_R

pub fn variant(&self) -> Variant<u8, RAMCFG_A>

Get enumerated values variant

pub fn is_ret(&self) -> bool

Checks if the value of the field is RET

pub fn is_partial(&self) -> bool

Checks if the value of the field is PARTIAL

pub fn is_off(&self) -> bool

Checks if the value of the field is OFF