Type Definition atsamd51p::qspi::ctrlb::DATALEN_R

type DATALEN_R = R<u8, DATALEN_A>;

Reader of field DATALEN

Implementations

impl DATALEN_R

pub fn variant(&self) -> Variant<u8, DATALEN_A>

Get enumerated values variant

pub fn is_8bits(&self) -> bool

Checks if the value of the field is _8BITS

pub fn is_9bits(&self) -> bool

Checks if the value of the field is _9BITS

pub fn is_10bits(&self) -> bool

Checks if the value of the field is _10BITS

pub fn is_11bits(&self) -> bool

Checks if the value of the field is _11BITS

pub fn is_12bits(&self) -> bool

Checks if the value of the field is _12BITS

pub fn is_13bits(&self) -> bool

Checks if the value of the field is _13BITS

pub fn is_14bits(&self) -> bool

Checks if the value of the field is _14BITS

pub fn is_15bits(&self) -> bool

Checks if the value of the field is _15BITS

pub fn is_16bits(&self) -> bool

Checks if the value of the field is _16BITS