atsamd51p::sdhc0::dbgr
Type Alias NidbgW

pub type NidbgW<'a, REG> = BitWriter<'a, REG, Nidbgselect>;

Field NIDBG writer - Non-intrusive debug enable

Aliased Type

struct NidbgW<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> NidbgW<'a, REG>
where REG: Writable + RegisterSpec,

pub fn idbg(self) -> &'a mut W<REG>

Debugging is intrusive (reads of BDPR from debugger are considered and increment the internal buffer pointer)

pub fn nidbg(self) -> &'a mut W<REG>

Debugging is not intrusive (reads of BDPR from debugger are discarded and do not increment the internal buffer pointer)