Module atsamd51p::sdhc0::feraces

Force Event for Auto CMD Error Status

Structs

ACMD12NE_W

Field ACMD12NE writer - Force Event for Auto CMD12 Not Executed

ACMDCRC_W

Field ACMDCRC writer - Force Event for Auto CMD CRC Error

ACMDEND_W

Field ACMDEND writer - Force Event for Auto CMD End Bit Error

ACMDIDX_W

Field ACMDIDX writer - Force Event for Auto CMD Index Error

ACMDTEO_W

Field ACMDTEO writer - Force Event for Auto CMD Timeout Error

CMDNI_W

Field CMDNI writer - Force Event for Command Not Issued By Auto CMD12 Error

FERACES_SPEC

Force Event for Auto CMD Error Status

W

Register FERACES writer

Enums

ACMD12NE_AW

Force Event for Auto CMD12 Not Executed

ACMDCRC_AW

Force Event for Auto CMD CRC Error

ACMDEND_AW

Force Event for Auto CMD End Bit Error

ACMDIDX_AW

Force Event for Auto CMD Index Error

ACMDTEO_AW

Force Event for Auto CMD Timeout Error

CMDNI_AW

Force Event for Command Not Issued By Auto CMD12 Error