Type Definition atsamd51p::sdhc0::pcr::SDBVSEL_R

type SDBVSEL_R = R<u8, SDBVSEL_A>;

Reader of field SDBVSEL

Implementations

impl SDBVSEL_R

pub fn variant(&self) -> Variant<u8, SDBVSEL_A>

Get enumerated values variant

pub fn is_1v8(&self) -> bool

Checks if the value of the field is _1V8

pub fn is_3v0(&self) -> bool

Checks if the value of the field is _3V0

pub fn is_3v3(&self) -> bool

Checks if the value of the field is _3V3