Type Definition atsamd51p::sercom0::usart_int::ctrla::TXPO_R

type TXPO_R = R<u8, TXPO_A>;

Reader of field TXPO

Implementations

impl TXPO_R

pub fn variant(&self) -> Variant<u8, TXPO_A>

Get enumerated values variant

pub fn is_pad0(&self) -> bool

Checks if the value of the field is PAD0

pub fn is_pad3(&self) -> bool

Checks if the value of the field is PAD3