atsamd51p::system_control
Module icsr

Interrupt Control and State Register

Structs

IcsrSpec

Interrupt Control and State Register

Enums

Nmipendsetselect

NMI set-pending bit

Pendstclrselect

SysTick clear-pending bit

Pendstsetselect

SysTick set-pending bit

Pendsvclrselect

PendSV clear-pending bit

Pendsvsetselect

PendSV set-pending bit

Type Aliases

IsrpendingR

Field ISRPENDING reader - Interrupt pending flag

IsrpendingW

Field ISRPENDING writer - Interrupt pending flag

IsrpreemptR

Field ISRPREEMPT reader - Debug only

IsrpreemptW

Field ISRPREEMPT writer - Debug only

NmipendsetR

Field NMIPENDSET reader - NMI set-pending bit

NmipendsetW

Field NMIPENDSET writer - NMI set-pending bit

PendstclrR

Field PENDSTCLR reader - SysTick clear-pending bit

PendstclrW

Field PENDSTCLR writer - SysTick clear-pending bit

PendstsetR

Field PENDSTSET reader - SysTick set-pending bit

PendstsetW

Field PENDSTSET writer - SysTick set-pending bit

PendsvclrR

Field PENDSVCLR reader - PendSV clear-pending bit

PendsvclrW

Field PENDSVCLR writer - PendSV clear-pending bit

PendsvsetR

Field PENDSVSET reader - PendSV set-pending bit

PendsvsetW

Field PENDSVSET writer - PendSV set-pending bit

R

Register ICSR reader

RettobaseR

Field RETTOBASE reader - No preempted active exceptions to execute

RettobaseW

Field RETTOBASE writer - No preempted active exceptions to execute

VectactiveR

Field VECTACTIVE reader - Active exception number

VectactiveW

Field VECTACTIVE writer - Active exception number

VectpendingR

Field VECTPENDING reader - Exception number of the highest priority pending enabled exception

VectpendingW

Field VECTPENDING writer - Exception number of the highest priority pending enabled exception

W

Register ICSR writer