Type Definition atsamd51p::tc0::count16::evctrl::EVACT_R

type EVACT_R = R<u8, EVACT_A>;

Reader of field EVACT

Implementations

impl EVACT_R

pub fn variant(&self) -> EVACT_A

Get enumerated values variant

pub fn is_off(&self) -> bool

Checks if the value of the field is OFF

pub fn is_retrigger(&self) -> bool

Checks if the value of the field is RETRIGGER

pub fn is_count(&self) -> bool

Checks if the value of the field is COUNT

pub fn is_start(&self) -> bool

Checks if the value of the field is START

pub fn is_stamp(&self) -> bool

Checks if the value of the field is STAMP

pub fn is_ppw(&self) -> bool

Checks if the value of the field is PPW

pub fn is_pwp(&self) -> bool

Checks if the value of the field is PWP

pub fn is_pw(&self) -> bool

Checks if the value of the field is PW