atsamd51p::tc0::count8::ctrla

Type Alias PrescsyncW

pub type PrescsyncW<'a, REG> = FieldWriter<'a, REG, 2, Prescsyncselect>;

Field PRESCSYNC writer - Prescaler and Counter Synchronization

Aliased Type

struct PrescsyncW<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> PrescsyncW<'a, REG>
where REG: Writable + RegisterSpec, REG::Ux: From<u8>,

pub fn gclk(self) -> &'a mut W<REG>

Reload or reset the counter on next generic clock

pub fn presc(self) -> &'a mut W<REG>

Reload or reset the counter on next prescaler clock

pub fn resync(self) -> &'a mut W<REG>

Reload or reset the counter on next generic clock and reset the prescaler counter