atsamd51p::tcc0::fctrlb

Type Alias BlankW

pub type BlankW<'a, REG> = FieldWriter<'a, REG, 2, Blankselect, Safe>;

Field BLANK writer - Fault B Blanking Mode

Aliased Type

struct BlankW<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> BlankW<'a, REG>
where REG: Writable + RegisterSpec, REG::Ux: From<u8>,

pub fn start(self) -> &'a mut W<REG>

Blanking applied from start of the ramp

pub fn rise(self) -> &'a mut W<REG>

Blanking applied from rising edge of the output waveform

pub fn fall(self) -> &'a mut W<REG>

Blanking applied from falling edge of the output waveform

pub fn both(self) -> &'a mut W<REG>

Blanking applied from each toggle of the output waveform