Type Definition atsamd51p::tcc0::fctrlb::HALT_R

type HALT_R = R<u8, HALT_A>;

Reader of field HALT

Implementations

impl HALT_R

pub fn variant(&self) -> HALT_A

Get enumerated values variant

pub fn is_disable(&self) -> bool

Checks if the value of the field is DISABLE

pub fn is_hw(&self) -> bool

Checks if the value of the field is HW

pub fn is_sw(&self) -> bool

Checks if the value of the field is SW

pub fn is_nr(&self) -> bool

Checks if the value of the field is NR