Type Definition atsamd51p::usb::host::fsmstatus::FSMSTATE_R

type FSMSTATE_R = R<u8, FSMSTATE_A>;

Reader of field FSMSTATE

Implementations

impl FSMSTATE_R

pub fn variant(&self) -> Variant<u8, FSMSTATE_A>

Get enumerated values variant

pub fn is_off(&self) -> bool

Checks if the value of the field is OFF

pub fn is_on(&self) -> bool

Checks if the value of the field is ON

pub fn is_suspend(&self) -> bool

Checks if the value of the field is SUSPEND

pub fn is_sleep(&self) -> bool

Checks if the value of the field is SLEEP

pub fn is_dnresume(&self) -> bool

Checks if the value of the field is DNRESUME

pub fn is_upresume(&self) -> bool

Checks if the value of the field is UPRESUME

pub fn is_reset(&self) -> bool

Checks if the value of the field is RESET