Crate atsamd51p

Peripheral access API for ATSAMD51P microcontrollers (generated using svd2rust v0.33.5 ( ))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Re-exports

pub use self::adc0 as adc1;

pub use self::sdhc0 as sdhc1;

pub use self::sercom0 as sercom1;

pub use self::sercom0 as sercom2;

pub use self::sercom0 as sercom3;

pub use self::sercom0 as sercom4;

pub use self::sercom0 as sercom5;

pub use self::sercom0 as sercom6;

pub use self::sercom0 as sercom7;

pub use self::tc0 as tc1;

pub use self::tc0 as tc2;

pub use self::tc0 as tc3;

pub use self::tc0 as tc4;

pub use self::tc0 as tc5;

pub use self::tc0 as tc6;

pub use self::tc0 as tc7;

pub use self::tcc0 as tcc1;

pub use self::tcc0 as tcc2;

pub use self::tcc0 as tcc3;

pub use self::tcc0 as tcc4;

Modules

ac

Analog Comparators

adc0

Analog Digital Converter

aes

Advanced Encryption Standard

ccl

Configurable Custom Logic

cmcc

Cortex M Cache Controller

core_debug

Core Debug Register

dac

Digital-to-Analog Converter

dmac

Direct Memory Access Controller

dsu

Device Service Unit

eic

External Interrupt Controller

etm

Embedded Trace Macrocell

evsys

Event System Interface

freqm

Frequency Meter

gclk

Generic Clock Generator

generic

Common register and bit access and modify traits

hmatrix

HSB Matrix

i2s

Inter-IC Sound Interface

icm

Integrity Check Monitor

mclk

Main Clock

nvmctrl

Non-Volatile Memory Controller

osc32kctrl

32kHz Oscillators Control

oscctrl

Oscillators Control

pac

Peripheral Access Controller

pcc

Parallel Capture Controller

pdec

Quadrature Decodeur

pm

Power Manager

port

Port Module

qspi

Quad SPI interface

ramecc

RAM ECC

rstc

Reset Controller

rtc

Real-Time Counter

sdhc0

SD/MMC Host Controller

sercom0

Serial Communication Interface

supc

Supply Controller

sys_tick

System timer

system_control

System Control Registers

tc0

Basic Timer Counter

tcc0

Timer Counter Control

trng

True Random Generator

usb

Universal Serial Bus

wdt

Watchdog Timer

Structs

Ac

Analog Comparators

Adc0

Analog Digital Converter

Adc1

Analog Digital Converter

Aes

Advanced Encryption Standard

CBP

Cache and branch predictor maintenance operations

CPUID

CPUID

Ccl

Configurable Custom Logic

Cmcc

Cortex M Cache Controller

CoreDebug

Core Debug Register

CorePeripherals

Core peripherals

DCB

Debug Control Block

DWT

Data Watchpoint and Trace unit

Dac

Digital-to-Analog Converter

Dmac

Direct Memory Access Controller

Dsu

Device Service Unit

Eic

External Interrupt Controller

Etm

Embedded Trace Macrocell

Evsys

Event System Interface

FPB

Flash Patch and Breakpoint unit

FPU

Floating Point Unit

Freqm

Frequency Meter

Gclk

Generic Clock Generator

Hmatrix

HSB Matrix

I2s

Inter-IC Sound Interface

ITM

Instrumentation Trace Macrocell

Icm

Integrity Check Monitor

MPU

Memory Protection Unit

Mclk

Main Clock

NVIC

Nested Vector Interrupt Controller

Nvmctrl

Non-Volatile Memory Controller

Osc32kctrl

32kHz Oscillators Control

Oscctrl

Oscillators Control

Pac

Peripheral Access Controller

Pcc

Parallel Capture Controller

Pdec

Quadrature Decodeur

Peripherals

All the peripherals.

Pm

Power Manager

Port

Port Module

Qspi

Quad SPI interface

Ramecc

RAM ECC

Rstc

Reset Controller

Rtc

Real-Time Counter

SCB

System Control Block

SYST

SysTick: System Timer

Sdhc0

SD/MMC Host Controller

Sdhc1

SD/MMC Host Controller

Sercom0

Serial Communication Interface

Sercom1

Serial Communication Interface

Sercom2

Serial Communication Interface

Sercom3

Serial Communication Interface

Sercom4

Serial Communication Interface

Sercom5

Serial Communication Interface

Sercom6

Serial Communication Interface

Sercom7

Serial Communication Interface

Supc

Supply Controller

SysTick

System timer

SystemControl

System Control Registers

TPIU

Trace Port Interface Unit

Tc0

Basic Timer Counter

Tc1

Basic Timer Counter

Tc2

Basic Timer Counter

Tc3

Basic Timer Counter

Tc4

Basic Timer Counter

Tc5

Basic Timer Counter

Tc6

Basic Timer Counter

Tc7

Basic Timer Counter

Tcc0

Timer Counter Control

Tcc1

Timer Counter Control

Tcc2

Timer Counter Control

Tcc3

Timer Counter Control

Tcc4

Timer Counter Control

Trng

True Random Generator

Usb

Universal Serial Bus

Wdt

Watchdog Timer

Enums

Interrupt

Enumeration of all the interrupts.

Constants

NVIC_PRIO_BITS

Number available in the NVIC for configuring priority